LightGBM inference result matches input type (#2129)
This makes the LightGBM inference return the same as the input type (either
float32 or float64) rather than always returning float64 per the API.

In addition, it adds a bit nicer handling for creation from a Buffer in that it
will also accept either a FloatBuffer or a DoubleBuffer.

+++ ml/lightgbm/src/test/java/ai/djl/ml/lightgbm/LgbmModelTest.java
@@ -64,7 +64,7 @@ public class LgbmModelTest {
                 floats.rewind();
                 NDArray array = manager.create(data, new Shape(10, 4), DataType.FLOAT32);
                 NDList output = predictor.predict(new NDList(array));
-                Assert.assertEquals(output.singletonOrThrow().getDataType(), DataType.FLOAT64);
+                Assert.assertEquals(output.singletonOrThrow().getDataType(), DataType.FLOAT32);
                 Assert.assertEquals(output.singletonOrThrow().getShape().size(), 10);
             }
         }
